requirements.txt
google-generativeai
langchain
pypdf
pandas
matplotlib
tiktoken
textract
transformers
openai
faiss-cpu
streamlit